pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cc-mapf"
version = "0.1.0"
description = "Connectivity-constrained multi-agent path finding research package"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "imageio>=2.34",
  "matplotlib>=3.8",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
ccmapf = "cc_mapf.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
